GH-2 : Implemented basic functionality
I've implemented the basic functionality of the test. It will now start a 1 minute countdown when you start typing into the textfield. After which it will calculate your results and display them. I've collapsed all the text down to a single line, removing new lines and tabs. This is to prevent false errors since these characters don't show up on the page and will be hard to enter into the text field.

I've fallen behind on the unit tests, partially because the hardcoded url to the text is not the same path when run in the test file. Also the styling is is in dire need of attention and more functionality should be added to improve the expereience.

Closes GH-2