GH-5 : added cancel function
Added a public cancel function that will stop the timer and call a reset function. The reset function empties the text from the input field and makes sure the event binding is correct. The cancel function is automatically called if the textfield loses focus. This may have to be rethought later.

Closes GH-5