GH-10, GH-11: Improved score notification
Changed the calculateWPM function to be calculateWPMStats. It now returns info about the wpm, adjustwpm, and errors. This is passed along to the notification function, which has been moved out. A default notification function is provided, which displays all the stats in an alert. This can be overridden in the defaults.

Closes GH-10, GH-11