GH-14 : refactored last listener and fetchText
Added a finish function so that I could pull out the afterTimeFinished listener and put it into the defaults. This allowed me to drop the addListeners function completely. Also refactored the fetchText. There is now a public fetchText method that calls the getText invoker. This will allow for the implementation of fetching the text to be changed if needed. It also allows a user to load text from outside of the component.